Extract ConfigService from MessageDB (2 functions, ~355 lines)
